Added multiple pages from landing to store
