+++ package.json
@@ -17,6 +17,7 @@
         "react-dnd": "^15.1.2",
         "react-dnd-html5-backend": "^15.1.3",
         "react-dom": "^17.0.2",
+        "react-router-dom": "^6.11.0",
         "react-scripts": "5.0.0",
         "ts-node": "^10.9.1",
         "typescript": "^4.5.5",
