+++ src/App.css
@@ -1,3 +1,7 @@
+a {
+    text-decoration: none;
+}
+
 .App {
     text-align: center;
     background-color: azure;
@@ -9,31 +13,28 @@
     pointer-events: none;
 }
 
-@media (prefers-reduced-motion: no-preference) {
-    .App-logo {
-        animation: App-logo-spin infinite 20s linear;
-    }
-}
-
-.App-header {
+.Intro-header {
     width: 100%;
     background-color: #282c34;
-    min-height: 40vh;
+    min-height: 100vh;
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
-    font-size: calc(10px + 2vmin);
+    font-size: calc(30px + 2vmin);
     color: white;
     background-image: url(https://img.freepik.com/premium-photo/vegetables-fruits-are-food-background-header-healthy-food-website_114160-2135.jpg?w=2000);
 }
+
 .description {
     border: 3px solid #969696;
 }
+
 .image {
     background-image: url(https://img.freepik.com/premium-photo/vegetables-fruits-are-food-background-header-healthy-food-website_114160-2135.jpg?w=2000);
     height: 250px;
 }
+
 .App-link {
     color: #61dafb;
 }
@@ -100,23 +101,18 @@
     background-color: ghostwhite;
 }
 
-@keyframes App-logo-spin {
-    from {
-        transform: rotate(0deg);
-    }
-    to {
-        transform: rotate(360deg);
-    }
-}
-
 .Trash-Can {
     display: flex;
 }
 
 .scroll-down {
     color: black;
+    text-decoration: none;
+    background-color: white;
+    font-weight: bolder;
 }
 .scroll-down:hover {
     color: black;
     background-color: orange;
+    font-size: larger;
 }
