+++ src/components/IntroHeader.tsx
@@ -1,17 +1,20 @@
 import React from "react";
 import "../App.css";
+import { Link } from "react-router-dom";
+
 export function IntroHeader() {
     return (
         <div>
-            <header className="App-header">
-                Welcome to the best grocery store on the planet
-                <a
-                    href="#middle"
-                    className="scroll-down"
-                    style={{ border: "5px solid Black" }}
-                >
-                    Shop Now
-                </a>
+            <header className="Intro-header">
+                <strong>Welcome to the BEST grocery store on the planet</strong>
+                <Link to={"/store"}>
+                    <span
+                        className="scroll-down"
+                        style={{ border: "10px solid Black" }}
+                    >
+                        SHOP NOW
+                    </span>
+                </Link>
             </header>
         </div>
     );
