+++ src/index.tsx
@@ -5,10 +5,13 @@ import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
+import { HashRouter } from "react-router-dom";
 
 ReactDOM.render(
     <React.StrictMode>
-        <App />
+        <HashRouter>
+            <App />
+        </HashRouter>
     </React.StrictMode>,
     document.getElementById("root")
 );
