+++ src/pages/Landing.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { IntroHeader } from "../components/IntroHeader";
+
+function Landing(): JSX.Element {
+    return (
+        <div>
+            <IntroHeader></IntroHeader>
+        </div>
+    );
+}
+
+export default Landing;
